+++ CHANGELOG.rst
@@ -4,6 +4,10 @@ Change log
 `Next version`_
 ~~~~~~~~~~~~~~~
 
+- Made file model mixins determine themselves whether they can accept an
+  upload or not.
+
+
 `0.3`_ (2017-06-21)
 ~~~~~~~~~~~~~~~~~~~